backend/requirements.txt
flask
numpy
pillow
matplotlib
kubernetes
flask-cors
